content-service/requirements.txt
pylint==2.8.3
Flask==1.1.2
Cerberus==1.3.2
flask_accept==0.0.6
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
python-slugify==5.0.2
psycopg2-binary==2.9.1